[#17] refactor: 프로필페이지 탭 컴포넌트 분리
