+++ src/components/profile/TabContent.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import TabPanel from "./TabPanel";
+
+const TabContent = () => {
+  return (
+    <>
+      <TabPanel value="pin" />
+      <TabPanel value="board" />
+    </>
+  );
+};
+
+export default TabContent;
